bolt | clean up dependencies
